FieldVerification functionnal

+++ src/Controllers/ChapterController.php
@@ -2,7 +2,6 @@
 
 namespace App\src\Controllers;
 
-// use App\src\Services\VerificationHelper;
 use App\src\Core\Controller;
 use App\src\Managers\ChapterManager;
 use App\src\Managers\CommentManager;
@@ -43,34 +42,37 @@ class ChapterController extends Controller
      */
     public function createChapter($isAdmin)
     {
-        unset($_SESSION['addSuccessMsg']);
-        unset($_SESSION['addErrorMsg']);
-        if (isset($_POST['save'])) {
-            $errors = FormVerificationHelper::checkField($_POST);
-            var_dump($errors);
-            if($errors){
-                $_SESSION['addErrorMsg'] = $errors;
+        if($isAdmin) {;
+            if (isset($_POST['save'])) {
+                $errors = FormVerificationHelper::checkField($_POST);
                 var_dump($errors);
-            }
-            if (!empty(htmlspecialchars($_POST['author'])) and !empty(htmlspecialchars($_POST['title'])) and !empty(htmlspecialchars($_POST['imgUrl'])) and !empty(htmlspecialchars($_POST['content']))) {
-                $newChapter = new Chapter($_POST);
-                $newChapter->setCreateDate(date("Y-m-d H:i:s"));
-                $newChapter->setUpdateDate(date("Y-m-d H:i:s"));
-                $newChapter->setPublished(false);
-                var_dump($insertLines);
-                if ($insertLines == false) {
-                    $insertLines = $this->chapterManager->insertInto($newChapter);
-                    $_SESSION['addSuccessMsg'] = 'Le nouveau chapitre à été enregistré';
-                    header('Location: /adminView');
-                } else {
-                    throw new Exception("Impossible de sauvegarder le chapitre");
+                if($errors){
+                    $_SESSION['addErrorMsg'] = $errors;
+                    var_dump($errors);
                 }
-            } else {
-                $_SESSION['addErrorMsg'] = 'Veuillez remplir tous les champs';
-                header('Location: /createChapter');
+                if (!empty(htmlspecialchars($_POST['author'])) and !empty(htmlspecialchars($_POST['title'])) and !empty(htmlspecialchars($_POST['imgUrl'])) and !empty(htmlspecialchars($_POST['content']))) {
+                    $newChapter = new Chapter($_POST);
+                    $newChapter->setCreateDate(date('Y-m-d H:i:s'));
+                    $newChapter->setUpdateDate(date('Y-m-d H:i:s'));
+                    $newChapter->setPublished(false);
+                    var_dump($insertLines);
+                    if ($insertLines == false) {
+                        $insertLines = $this->chapterManager->insertInto($newChapter);
+                        $_SESSION['addSuccessMsg'] = 'Le nouveau chapitre à été enregistré';
+                        header('Location: /adminView');
+                    } else {
+                        throw new Exception('Impossible de sauvegarder le chapitre');
+                    }
+                } else {
+                    $_SESSION['addErrorMsg'] = 'Veuillez remplir tous les champs';
+                    header('Location: /createChapter');
+                }
             }
+            echo $this->render('adding_chapter.html.twig', ['isAdmin' => $isAdmin]);
+        } else {
+            throw new Exception ('Page réservé à l\'administration !');
         }
-        echo $this->render('adding_chapter.html.twig', ['isAdmin' => $isAdmin]);
+
     }
 
     /**
@@ -87,7 +89,6 @@ class ChapterController extends Controller
             ['uniqueChapter' => $this->chapterManager->findOneBy(array('id' => $chapterId)),
             'commentList' => $this->commentManager->findBy(array('chapterId' => $chapterId), array('createdDate' => 'ASC')),
             'totalReporting' => $this->reportingManager->totalReportCount(),
-            'chapterNumber' => $chapterId,
             'isAdmin' => $isAdmin]
         );
     }
@@ -102,11 +103,15 @@ class ChapterController extends Controller
      */
     public function updateChapter($chapterId, $isAdmin)
     {
-        echo $this->render(
-            'modifying_chapter.html.twig',
-            ['uniqueChapter' => $this->chapterManager->findOneBy(array('id' => $chapterId)),
-            'isAdmin' => $isAdmin]
-        );
+        if($isAdmin) {
+            echo $this->render(
+                'modifying_chapter.html.twig',
+                ['uniqueChapter' => $this->chapterManager->findOneBy(array('id' => $chapterId)),
+                'isAdmin' => $isAdmin]
+            );
+        } else {
+            throw new Exception ('Page réservé à l\'administration !');
+        }
     }
 
     /**
@@ -115,17 +120,21 @@ class ChapterController extends Controller
      * @param mixed $chapterId
      * @throws Exception
      */
-    public function deleteChapter($chapterId)
+    public function deleteChapter($chapterId, $isAdmin)
     {
-        $deleteChapter = $this->chapterManager->findOneBy(array('id' => $chapterId));
-        $this->chapterManager->delete($deleteChapter);
-        $this->commentManager->deleteFrom($deleteChapter);
-        // if ($deleteLines == false) {
-        //     throw new Exception("Le chapitre " . $chapterId . " n'existe pas");
-        // } else {
-            $_SESSION['deleteMsg'] = 'Le chapitre ' . $chapterId . ' et ses commentaires ont bien été supprimés';
-            header('Location: /adminView');
-        // }
+        if($isAdmin) {
+            $deleteChapter = $this->chapterManager->findOneBy(array('id' => $chapterId));
+            $this->chapterManager->delete($deleteChapter);
+            $this->commentManager->deleteFrom($deleteChapter);
+            if ($deleteLines == false) {
+                throw new Exception('Le chapitre  ' . $chapterId . ' n\'existe pas');
+            } else {            
+                $_SESSION['deleteMsg'] = 'Le chapitre ' . $chapterId . ' et ses commentaires ont bien été supprimés';
+                header('Location: /adminView');
+            }
+        } else {
+            throw new Exception ('Page réservé à l\'administration !');
+        }
     }
 
     /**
@@ -134,12 +143,16 @@ class ChapterController extends Controller
      * @param mixed $chapterId
      * @throws Exception
      */
-    public function publishChapter($chapterId)
+    public function publishChapter($chapterId, $isAdmin)
     {
-        $publishChapter = $this->chapterManager->findOneBy(array('id' => $chapterId));
-        $publishChapter->setPublished(true);
-        $this->chapterManager->update($publishChapter);
-        header('Location: /adminView');
+        if($isAdmin) {
+            $publishChapter = $this->chapterManager->findOneBy(array('id' => $chapterId));
+            $publishChapter->setPublished(true);
+            $this->chapterManager->update($publishChapter);
+            header('Location: /adminView');
+        } else {
+            throw new Exception ('Page réservé à l\'administration !');
+        }
     }
 
     /**
@@ -149,28 +162,31 @@ class ChapterController extends Controller
      * @return void
      * @throws Exception
      */
-    public function updateChapterAction($chapterId)
+    public function updateChapterAction($chapterId, $isAdmin)
     {
-        $_SESSION['modifySuccessMsg'] = '';
-        if (isset($_POST['save'])) {
-            if (!empty(htmlspecialchars($_POST['chapterTitle'])) and !empty(htmlspecialchars($_POST['chapterImg'])) and !empty(htmlspecialchars($_POST['chapterImg']))) {
-                $updatedChapter = $this->chapterManager->findOneBy(array('id' => $chapterId));
-                $updatedChapter->setTitle(htmlspecialchars($_POST['chapterTitle']));
-                $updatedChapter->setImgUrl(htmlspecialchars($_POST['chapterImg']));
-                $updatedChapter->setContent(htmlspecialchars($_POST['chapterContent']));
-                $updatedChapter->setUpdateDate(date("Y-m-d H:i:s"));
-                $updatedChapter->setPublished(true);
-                $updateLines = $this->chapterManager->update($updatedChapter);
-                if ($updateLines == false) {
-                    throw new Exception("Impossible de mettre à jour le chapitre");
+        if($isAdmin) {
+            if (isset($_POST['save'])) {
+                if (!empty(htmlspecialchars($_POST['chapterTitle'])) and !empty(htmlspecialchars($_POST['chapterImg'])) and !empty(htmlspecialchars($_POST['chapterImg']))) {
+                    $updatedChapter = $this->chapterManager->findOneBy(array('id' => $chapterId));
+                    $updatedChapter->setTitle(htmlspecialchars($_POST['chapterTitle']));
+                    $updatedChapter->setImgUrl(htmlspecialchars($_POST['chapterImg']));
+                    $updatedChapter->setContent(htmlspecialchars($_POST['chapterContent']));
+                    $updatedChapter->setUpdateDate(date('Y-m-d H:i:s'));
+                    $updatedChapter->setPublished(true);
+                    $updateLines = $this->chapterManager->update($updatedChapter);
+                    if ($updateLines == false) {
+                        throw new Exception('Impossible de mettre à jour le chapitre');
+                    } else {
+                        $_SESSION['modifySuccessMsg'] = 'Chapitre modifié et enregistré';
+                        header('Location: /readChapter/' . $chapterId);
+                    }
                 } else {
-                    $_SESSION['modifySuccessMsg'] = 'Chapitre modifié et enregistré';
-                    header('Location: /readChapter/' . $chapterId);
+                    $_SESSION['error'] = 'Veuillez remplir tous les champs';
+                    header('Location: /updateChapter');
                 }
-            } else {
-                $_SESSION['error'] = 'Veuillez remplir tous les champs';
-                header('Location: /updateChapter');
             }
+        } else {
+            throw new Exception ('Page réservé à l\'administration !');
         }
     }
 }
